Split large files: NavGraph, SettingsScreen, Database
Remove duplicate route/entry/component definitions from NavGraph (1098→283 lines),
SettingsScreen (1033→167 lines), and CivitDeckDatabase (892→213 lines) by completing
the extraction into separate files that already existed but were never committed.

New files for settings: AdvancedComponents, BehaviorComponents, FeedComponents,
NotificationComponents, StorageComponents, UpdateComponents.
New files for database: migrations/Migrations1to10, 11to20, 21to30, 31to38, SeedData.

Closes #617
